Initial commit
This project has grown in size and complexity to the point that I'd
be annoyed if I accidentally deleted it, so it is now being tracked
by Git and pushed off site.

It is still too much of a disorganized prototype to try to properly
track atomic changes in Git commits, so I am just going to dump any
notes and the current status of things in the README file for
better visibility for now.