Support cross-distro SELinux labeling everywhere
A side effect of this is that an ext4 image is always produced now,
so that the target system's packages can be installed onto a file
system that can be booted and modified.  This is probably for the
best regardless, since it ensures ACLs, extended attributes, and
other such metadata are captured in squashfs images regardless of
the host machine's file system choice.