Disable logins by default
Require the system definition file to grant access as appropriate.
For testing, an option was added to create user accounts with no
password (since this shouldn't be encouraging people to write
passwords on the command line, and SSH disables empty passwords
anyway, so it can be set as needed when logged into the system).

Also make the container's PS4 level color blue instead of red, so
it looks more like "info" rather than "error" at a glance.