Fix Gentoo a bit
This will auto-detect the latest stage3 autobuild now so that old
commits here will still be usable after old autobuilds are deleted
from the mirrors.  The pointer to the latest autobuild has no
cryptographic signature to be verified, though.  Since it is only
used to read a timestamp of an autobuild that is cryptographically
verified, the worst that can happen from this is a downgrade attack
(which could already happen if the server is compromised since old
autobuilds still have valid signatures with the same key).  The
software in the stage3 will always be upgraded to current before
building anything anyway.