Move individual package configuration steps into dropin files
Having all the different configuration files written by the same
function was starting to get unreadable, so hopefully this will
make it easier to find things.

This also starts automating generating caches/databases for any
installed packages that support it, which allows images to be
customized by just writing modules into place and letting the build
system handle the rest.  This will probably need to moved to dropin
files as well as it continues to be expanded.