Banish CentOS 7 workarounds to their own file
This tests building the squashfs image in the relabeling VM to fix
some edge cases of SELinux on squashfs using CentOS 7.  It plumbs
through /dev/kvm from the host to the build root now so it can be
done efficiently.

It also drops the iptables option in favor of making it another
automatically-preconfigure-if-installed package.