Add CentOS 8 support
CentOS 8 still does not support networkd, so NetworkManager is the
required package for networking.  It still does not support busybox
for the relabeling initrd.  CentOS 8 dropped ImageMagick, so it
can't build the UEFI splash screen anymore.  Its kernel remains too
old to use dm-init, so verity continues to be handled in userspace.

CentOS 7 support is still included, but it has been moved to its
own file since it requires much special handling of ancient package
versions.  The CentOS 7 file can't be used directly; the regular
CentOS file reads it when the release version is set to 7.  There
is some nice code reuse now via the Fedora -> CentOS 8 -> CentOS 7
inheritance, so it's not too bad to keep alive.