Support creating executable virtual machine images
This adds the ability to create an ESP on executable disk images
like the containers use, and it renames the option from "nspawn" to
"executable" to show that any general script can be used to run the
image.  For a basic virtual machine, the build should create the
file "launch.sh" with the following command.

    exec qemu-kvm -nodefaults \
        -bios /usr/share/edk2/ovmf/OVMF_CODE.fd \
        -cpu host -m 4G -vga std -nic user \
        -drive file="${IMAGE:-disk.exe}",format=raw,media=disk "$@"

Without the "uefi" option, the image can only be started as a
container (not a VM) since making the image executable is
incompatible with non-UEFI booting.  However with the "bootable"
option, the image can be "booted" as an nspawn container since it
will still have a fully configured init system.  The "ramdisk" and
"uefi" options together will result in a disk image with only the
ESP (and no root file system), so it cannot be run as a container.
Here is a list of options and boot possibilities:

    no options:   Run as an app container
    bootable:     Run as an app container, or boot with nspawn
    uefi:         Run as an app container, boot with nspawn, or VM
    uefi+ramdisk: Run as a VM

This also works around the UEFI splash image colors by converting
32-bit depth to 24-bit with "-type truecolor", since the problem is
from systemd's pixel blending.