Switch UEFI splash screen workarounds
The real issue is that systemd ignores the channel bit masks, which
was being worked around by using 24-bit pixels with no alpha
channel to match what is expected.  This switches back to 32-bit
with alpha and reorders the channels using a color matrix so that
all colors are in the order that the boot stub expects, and it gets
the best results with alpha blending again.  The downside is that
the logo image will no longer be viewable on the host, but it works
correctly when the system boots, and that's what matters.