Add support for Secure Boot signing
This only supports file-based keys at the moment.  Also, it is bad
to have the private key available through the entire containerized
procedure.  This will need to be revised later to support smart
cards for signing anyway, but for the time being, this works well
for building and signing non-critical-security systems.