Push Gentoo a step in the right direction
Everything should be supported in Gentoo now, but it is set to only
run the targeted SELinux policy in permissive mode due to it not
being compatible with systemd upstream.

It will probably require workarounds to be added continually to the
profile(s) to support different package sets, but this seems to be
working for console-based systems so far.