Just build networkd for CentOS 8
I didn't want to rebuild components of a binary distro (especially
a core package like the init system), but I also don't want to be
held hostage by Red Hat's terrible decisions (hence the existence
of this project).  This will now build networkd into the systemd
package and install the custom version when the networkd option is
enabled for the target system, so a system that leaves the option
disabled and installs NetworkManager will have networking and still
use the unmodified systemd package.