Add an example Gentoo system
This is a starting point for building an image for my current
laptop system.  It is tuned to only run on that hardware to show
what Gentoo can do, since the Fedora and CentOS options are built
for compatibility to run on anything.

The file is far from complete.  Many kernel settings need to be
added, and this is intended to eventually have a modern desktop
environment.  It currently just has twm and xterm for a graphical
environment, with xrandr to reduce the ridiculous 4K resolution.