Move NVMe support to an option
The distros have to load optional code differently to support the
same functionality, so this seems like something that should be
abstracted rather than handled in each system's customizations.

The issue in Gentoo is just a matter of enabling NVMe support as a
builtin so a verity root can be used on such a device.  In Fedora,
the NVMe module is loaded by udev, but the dm-init setup happens
before the initrd starts running, so this adds an NVMe dependency
on the verity module.  CentOS kernels are too old for dm-init, so
there shouldn't be an issue, but it inherits the added dependency
from Fedora anyway to ensure things will work if anything changes.