Improve cross-compiling a bit
This makes an x86_64 build system default to a multilib-enabled
profile now, since there are apparently bad dependencies that
require 32-bit libraries installed on the host when cross-compiling
to a 32-bit target (even though the entire point of building a
sysroot is to not install target system packages on the host).  It
still uses the no-multilib profile for target x86_64 systems.