Commit aimless churn
Most of this is updates for fixes applied upstream and improving
cross-compiling support.

One significant change is that files from pstore are now logged to
the journal and freed.

The /etc Git overlay has been moved from a generator to a service
since it was abusing the purpose of generators, but this loses
support for configuring units and CentOS 7 in general (since it is
implemented with modern systemd features).  It might be adapted to
the necessary initrd implementation later.