Implement EROFS for uncompressed read-only images
This should be a nice efficiency boost in builds for weird old
systems, which are only targeted by Gentoo at the moment (since all
other supported distros are only for x86_64 now).  Nevertheless,
Arch also supports EROFS.  Fedora and CentOS are disappointments.

It's a bit early for this feature since none of the distros has an
erofs-utils package out of the box, and it has only been in a
released kernel for a few days.  (Arch is still on Linux 5.3, but
EROFS is enabled in staging.)  There is an overlayfs bug that is
patched for Gentoo, and Arch works around it by forcing the use of
xattrs even when SELinux is disabled.  The mkfs.erofs program does
not yet support excluding files, so it might be incompatible with
the current implementation of SELinux labeling.  (That's not an
issue yet since Gentoo doesn't support SELinux with systemd, so it
shouldn't be used on systems that care enough about efficiency to
not use squashfs, and Arch doesn't support SELinux at all.)

This also fixes the systemd UEFI boot stub splash screen images for
ImageMagick 7 in Arch and Gentoo.  It also drops bad loop device
juggling when creating an ESP for executable disk images in favor
of using mtools.