Add another hook for repo configuration
The initialize_buildroot() function can now be defined to be run
after the distro container image is extracted and configured, but
before any package managemer commands are run.  This allows systems
to configure repository settings or modify transaction behavior for
the initial update.

Gentoo specifically has more use for this, so its buildroot is now
built from the following ordered steps to support cross-compilation
in the native root (e.g. to make GRUB modules for the target).

  - Download/verify/extract the stage3.
  - Write the portage profiles.
  - Run initialize_buildroot() for initial profile changes.
  - Update the stage3 root packages / convert to systemd.
  - Install cross-compilers.
  - Install all requested buildroot packages.
  - Run customize_buildroot(), then proceed with the target build.

Additionally in Gentoo, most relevant bugs are now being noted and
patched instead of using upstream-compatible workarounds.