Make ARMv7-A default to hard float ABI
The Gentoo profile forces VFP, so make the same assumption.  Note
that the Chromebook example uses the defaults since it doesn't set
its own ABIs, so that system gets an upgrade from this.

This also moves some ugly patching code for certain packages to its
own optional function that can be shared between systems.  The big
package is Firefox.  I was doing all the patching and experimenting
in the Mac mini example, but Firefox does not support PowerPC
anymore.  (Most of it looks easily portable, but it also requires
implementing the ISA for WebAssembly etc., which I am not planning
to do myself.)  The Firefox build is broken for all the other
targets that do not have th same architecture as the build system,
since cbindgen mixes include paths between the host and sysroot.
The only Gentoo example that builds Firefox is therefore the laptop
system, which assumes that the build system is the target.