Add an ARM netbook as a cross-compiling example
This is just to have another weird CPU (pre-ARMv7) to target for
experimentally building stuff.  I didn't put much effort into the
hardware support, so I'm using it with USB input and booting from
the SD card.  It might require GPIO fiddling to enable components.

Also, networkd was added to EPEL, so CentOS 8 is switched to that
instead of building it since that's what everyone else is going to
be using.  It's from a newer version than the included systemd, so
it will have many more fixes and features, too.