Make the sample formatter produce a hybrid MBR
This will support all of the GPT-based systems as well as the old
WM8505 netbook's U-Boot firmware now.

Also, this adds a daemon to that netbook to power the integrated
USB wireless device.  Also also, it overrides the runtime Emacs
dumping service in favor building the dump file into the image with
a system-specific emulator call (since it is horribly slow to
generate on boot).

Note that everything works with the fresh Linux 5.6 kernel except
the NVIDIA drivers.  I'm going to just leave that package as is for
this kernel update, rather than patching and reverting when it's
fixed upstream again.  A compatible update is usually released
quickly, and it will be automatically used when available.