Add support for Fedora 32
Fedora 30 support isn't going to be removed yet.  It's technically
still supported upstream for a few more days, but there is no other
viable i686 binary distro to build the game containers.  The idea
was to move to openSUSE, but it's been broken for a few weeks due
to missing signatures for its container environment.  (Comment out
the "verify_distro" line to build it anyway; running an unsigned
distro will not be supported out of the box.)

This also includes some fiddling with RISC-V UEFI.  It builds the
OpenSBI/U-Boot firmware image to provide a UEFI environment in QEMU
and a patched GRUB binary as the default boot option.  The Linux
stub patches are not applied yet since there are segfaults.  The
systemd stub does not exist at all.  Therefore, the example RISC-V
system by default will continue to start the non-UEFI Linux kernel
as before, but it can switch to booting into UEFI GRUB by changing
which firmware file QEMU uses.