Implement verity signatures as a new optional security layer
A lot of changes have accumulated since I was waiting for fixes to
be included in new upstream container images.  I tend to get
refactory while waiting around.

Verity signature support is not enabled in binary distro kernels,
so it's currently only for Gentoo.  It requires adding a signature
to a user keyring in userspace, so it has to be done in an initrd
to mount root.  Maybe later I can add a dumb kernel module to
initialize a default user keyring or something like that to set up
the signature with no initrd.  The signature authenticates the
verity root hash for platforms that don't have Secure Boot.  Unlike
Secure Boot, it can authenticate more verity devices from untrusted
sources after booting.  The proposed IPE LSM policy can also add
restrictions based on signed verity devices.

The save_boot_files function has been reordered to run after verity
is finished, and the RAM disk initrd is generated before the final
kernel.  This can allow Gentoo to build the kernel with all other
resources available, so the command-line and an initramfs could be
built into the kernel.  This is essentially a replacement for the
UEFI stub on non-UEFI platforms to build everything into a single
file, greatly simplifying other integrity solutions.

The optional certificate and private key pair has been split into
three different pairs.  Keys can now be specified for Secure Boot,
verity signatures, and general signing individually.  When the
Secure Boot or verity keys are not specified, they will default to
using the general signing keys.  The -c and -k parameters are used
to specify the general signing keys on the command-line, while the
others need to be in the configuration file's options.  If general
signing keys are not provided, temporary keys are generated to be
used for one build then discarded.

Secure Boot is now always enabled with UEFI builds for simplicity,
but it is still gated on an option that can disable it.  When not
given a certificate and key, it will have a signature from the
temporary key which will be useless unless the one-time certificate
is added to the trusted Secure Boot keyring.  The unsigned UEFI
executable will still be available in any case.

The README has a new compatibility table for distros' major feature
support.  To sum up the whole section for picking a distro, if you
don't care about...
  - build time, configuration effort, or SELinux, then use Gentoo.
  - uncommon platforms, then use Fedora.
  - anything and hate Fedora and Gentoo, then use Arch or openSUSE.
  - anything and hate yourself, then use CentOS.

The Fedora 30 and CentOS 7 distro implementations have been moved
to a new legacy directory.  That's where obsolete stuff will remain
functional but might be dropped arbitrarily.  Today is also the
last day of support for Fedora 30, so it will no longer be updated.

CentOS 7 has fixed building 32-bit containers on 64-bit systems.

The /etc/os-release file is now saved for container builds, too.

EROFS utilities have been updated in all distros, so the exclude
path workarounds have been dropped.

The RISC-V example is now using UEFI unconditionally.  It's going
to be waiting on the systemd UEFI stub for a long time before it's
fully on par with other architectures, but it's close.