Migrate 32-bit example containers to openSUSE
Fedora 30 is unsupported now, so its i686 support is dead, but
coincidentally openSUSE has finally fixed container signatures, so
it is now a viable replacement.  The x86_64 containers are still
using Fedora 32, so all containers are on supported distros again.

Some 32-bit openSUSE programs fail to execute with older nspawn
versions, so in case the containers fail to build, the solution is
to upgrade systemd on the host.

Relatedly, this adds --pipe to nspawn calls to fix passing file
descriptors in systemd versions since 242.  The flag is only used
when detected, so everything is still compatible with 241.