Minor fixes for Linux 5.7 in Gentoo
This mostly updates the RISC-V example and NVIDIA drivers to work
better with Linux 5.7.

It also moves the example containers' setup scripts to a file at
/init in the actual container file system rather than passing it as
a here-document.  This frees up stdin for the user to be able to
pass data to the container programs, and it should allow dropping
the --pipe option for normal usage if desired.  Containers that do
not require a setup script just symlink the executable to /init so
there is a consistent entry point.

A quick prototype of switching to the proprietary NVIDIA drivers in
the game containers is implemented for Arx Fatalis.  It works the
same way for everything else, but it complicates things, so it is
only present in one game for now.