Add Ubuntu
So I've finally snapped and added something Debian-based.  I needed
a binary distro that supports verity root hash signatures, and
waiting for certain other kernel maintenance teams to respond is a
completely hopeless waste of time.

This implementation decompresses the kernel and inserts the verity
signing certificate into the binary, but that decompression loses
UEFI support.  The signature verification only accepts certificates
in the built-in keyring, not even the secondary trusted keyring or
platform (UEFI) keyring.  There is no provided way to recompress
and reattach the UEFI stub without the kernel build environment,
but maybe later I can rig up something simple to stuff the modified
vmlinux into the original vmlinuz.

Also note that this signed hash is loaded into the session keyring
instead of the user keyring as documented, because it happens in a
systemd service in the Ubuntu initrd.  That way, it works with the
default KeyringMode=private since systemd links the user keyring to
the session keyring here.

The example containers no longer carry the --pipe flag in their
nspawn calls.  It seems like a debug option since it isn't needed
for normal usage, so anyone that needs to pass file descriptors can
just edit the example scripts to add it back.

Gentoo's omnipresent churn is addressed as well, of course.  One
noteworthy change is the upstream Rust cross-compilation support,
which allowed dropping my implementation.  Since it's supported out
of the box and no longer patched in during Firefox build fixing,
any Rust project should now be usable with no modifications.