Implement a Gentoo option to build monolithic kernel images
This builds the kernel command-line and an initramfs (when one is
used) into the kernel image, and it ignores them when they provided
by the bootloader.  It is meant as a way to lock down boot-time
options into a single kernel file for any platform, the same way it
is handled by the systemd UEFI stub.  It should allow easier boot
integrity implementations where Secure Boot is not available.

The monolithic option is used on the ARM example systems now to
enable verity signatures as a lazy work around to avoid having to
add separate initrd handling to their U-Boot stuff.

Gentoo officially does not support split bin/sbin directories when
using UsrMerge, so the layout was changed to symlink sbin to bin.

A dumb retry loop was added when using verity with an initrd, since
some platforms can take a long time to recognize the device when
booting from USB, and rootwait has no benefit to veritysetup.

I've ported Firefox to non-SSE i686 for no reason and added it to
the fit-PC example system.  (I've also added WindowMaker as a real
desktop for it.)  It abuses Gentoo's new Rust cross-compiling setup
to build i586 libraries, since Rust believes that i686 implies that
SSE2 is available.  There is a big ugly patch to switch it to i586
Rust during an i686 build and to drop other assumptions throughout
Firefox that i686 means SSE2.