Build ramdisks without a disk image when possible
This is an optimization so the ramdisk option doesn't need to set
up a loop device and mount the root file system.  Instead, the
entire operating system installation is packed directly into the
cpio archive.

An initrd only needs to include a file system image in three cases:

  - With SquashFS, to keep the file system compressed in memory.
  - With verity, to check pregenerated hashes of the image.
  - With SELinux, to preserve extended attributes (the labels).

Basic read-only mode (without SquashFS or verity) is supported by
remounting root read-only after the kernel extracts the cpio into
it.  This could be seen as a slight loss of security over the
previous behavior of mounting an EROFS image (where the file system
driver does not support writing) since it could be remounted as
read-write again, but you should probably just enable verity if
immutability in the ramdisk matters that much.  (Mounting writable
overlays would cause the same problem in any case, so it's not that
big of a difference.)

SELinux could have been handled similarly, where the policy would
be loaded and the file system relabeled after the kernel extracts
the cpio (since the cpio format can't store extended attributes by
itself), but that is a weird operation to run on every boot that is
affected by the number of extracted files to relabel.  There is no
behavior change with this commit when using SELinux to avoid that.

There is a bit of a downgrade with this where early microcode
updates are not included in this format since the initrd was not
generated by dracut.  This is semi-intentional since Gentoo doesn't
use dracut and builds the microcode into the kernel, but this is a
regression for the binary distros.  I'll figure out a clean way to
handle this eventually, but until then, just enable one of the
options that use a file system image to keep the previous behavior.