Build more Firefoxes
This uses Gentoo's cross-Rust support to create the Thumb runtime
libraries for ARMv7-A as Gentoo wants to use for Firefox with NEON.
That allows building an unmodified Firefox for the Chromebook
example system.

All of the graphical Gentoo example systems were switched to
WindowMaker for consistency on the desktop.

I've reverted the verity signature and monolithic kernel options
from the netbook example, since the builtin initramfs seemed to
result in it having trouble booting.  I may be pushing that chip's
128MiB of RAM too far.