Enable support for verity signatures in Fedora
It's the same story as Ubuntu, where enabling verity signatures
removes UEFI support.  I've also lazily left it enabled for both
Fedora 32 and 31, but the kernel hasn't been rebased on 5.8 for
Fedora 31 yet, so it won't work there.  It will all magically work
again when they push the kernel rebase to the f31 branch.  Verity
signatures will never work for Fedora 30 since that branch is dead.

This also fixes Gentoo's recent pile of incompatibilities.  Lots of
upstream fixes and efficiency improvements are included.