Generalize GPT image generation
This basically renames the "executable" option to "gpt".  Now the
option will produce GPT images without making them executable by
default, only doing so when launch.sh exists.  This is to allow
creating full disk images that can be extended e.g. with repart.d.
A secondary option "esp_size" can be used to force a specific size
of the ESP in the GPT image since systemd-repart won't resize
existing partitions.

The implementation was left semi-extensible in case it is feasible
to use other formats like APM or DOS partitions interchangeably in
the future.  Upstream work is required to be able to use repart.d
provisioning with other partition formats.