Use repart.d in the Chromebook example
This runs systemd-repart as a demonstration of what it can do.  The
produced gpt.img file only contains two partitions: a 16MiB slot to
store a ChromeOS-style kernel and a partition holding the root file
system.  When the image is written to a microSD card and booted,
systemd-repart expands the root partition to 1GiB (to allow growth
with updates--the root FS is currently ~300MiB), adds another 16MiB
ChromeOS kernel, adds another 1GiB root FS, and allocates all
remaining disk space for persistent /var.  The /var partition is
automatically formatted as ext4 and mounted for immediate use, and
it supports factory-reset to start the system anew.

A new systemd generator was added as a workaround to specify the
target device to systemd-repart when the root file system is on
verity, until that is implemented upstream.

This update also adjusts some Gentoo Rust behavior.  Now the Rust
librsvg version is always blocked in the native build root and
always used while cross-compiling.  The netbook example explicitly
blocks the Rust version to save build time since nothing else uses
Rust on that system.  This allows using the more secure version
everywhere Rust is already available while not wasting hours of
build time elsewhere.  It also moves custom Rust target definitions
from the general Firefox fixes to the example systems with the
affected CPUs so it is more properly scoped.  Rust is also falling
back to using stable versions instead of testing since all of the
cross-compiling bits are stabilized.

All references to Fedora 30 have been dropped from the README.  The
functionality is still present in case 32-bit Fedora is needed for
something, but it is well past time to hide it since it has been
EOL for half a year now.