Automate cross-compiling Go in Gentoo
This predefines an appropriate GOARCH so that Go packages are built
for the correct target architecture as needed.  It also sets GOARM
for ARM targets to get the best FPU behavior.  Go is not available
for all architectures; half the CPUs I target are unsupported: no
ARMv4, no PowerPC, no SuperH, and Go 1.15 was the last version that
supported x86 without requiring SSE2 (i.e. no Geode).