Fix some Gentoo dependencies
This is a small update to keep Gentoo building.  They're playing
games with librsvg, so this makes it work in the native root again.

A pile of new testing packages is being accepted to fix cross-root
dependencies.  For most architectures, this just means that fewer
packages will be needlessly installed in the native root.  For the
multilib-aware targets, this means that architectural dependencies
are no longer being applied across roots, so it opens the door to
always using no-multilib native profiles and building multilib
target systems.  Most projects related to GNOME are still problems
as always, though, so they must be excluded for this at the moment.

These newly accepted testing packages also fix the broken build
dependency graph around X11, so an entire emerge call has been
eliminated.  (It is still called for SELinux, but my fix for that
has been applied in the upstream eclass, and the ebuilds should be
fixed to use it during the next policy release, so that will be
dropped soon as well.)

I've added a new Python script to query configured USE flags for
uninstalled packages since there are no tools to handle this.  It's
used for bootstrapping packages with features that create circular
dependencies, so now no unnecessary packages are built.  Support
for systemd with crypsetup has been added with this.