Fix Rust packages in Gentoo again
This is another update just to keep Gentoo building.  They now
require Rust for Adwaita icons.  The problem is only going to get
worse, so I've made all the example systems use the Rust librsvg
version.  Rust now has to be built when anything using GTK+ or SVG
is installed on the target system.

To try to remain efficient, the native librsvg build required by
UEFI images will still pick the old non-Rust version.  If any of
the target system packages require GTK+, it will end up in the
native build dependencies and force librsvg to be upgraded.  This
method should actually be faster than if the Rust librsvg version
was installed originally, since building it during buildroot setup
would take hours compiling LLVM and Rust with no other packages to
build in parallel.  That's the best I could do for now, given that
Gentoo is putting everyone in this terrible situation.  Rust really
needs to be built into the stage3 when basic stuff depends on it.

Some other random fixes:  Emacs fixes are all applied upstream, so
it can now be installed as a regular package with no manual fixing.
The policycoreutils dbus flag can now be respected due to the other
D-Bus and boostrapping fixes.  More unnecessary package builds are
being skipped to make it faster, which is handy now that everything
needs to waste time building Rust.