Push accumulating small fixes
Ubuntu and openSUSE no longer depend on a default cursor theme, so
explicitly pick Adwaita to allow GNOME to start again.  Fix home
directories' initial content on openSUSE.

Dracut initrds are now reproducibly built.

Fix the RISC-V UEFI patch to work with newer stable kernels.  Try
the new systemd profile as the default for RISC-V, and fix its mask
list to support verity setup in the ramdisk.

Build the fbdev video driver for experimenting on the Geode.  It
would be nice to be able to drop the Geode video driver and its
legacy requirements on /dev/mem etc.

Update Gentoo flags to match the new default Python version.

Gentoo changes have fixed even more package dependencies.  An extra
emerge call for SELinux was able to be dropped, so there are no
remaining workarounds for producing a functional dependency graph.