Make all example Gentoo builds use a graphical desktop by default
(That does not include the RISC-V system which is still headless.)

They now use LXDM for the login screen and Xfce for the desktop, so
they have a lightweight but modern GTK+ 3 GUI.

The netbook system is an exception, which uses XDM and WindowMaker
to skip installing GTK+.  LXDM and Xfce both build and run on that
platform, but its no-FPU ARMv5 CPU and 128MiB of RAM do not result
in an enjoyably responsive GTK+ 3 environment.

Gentoo dropped the EOL Linux 5.9 kernels, so systemd-repart is
broken on the Chromebook example again.  Use a preformatted GPT
disk to work around it until there is a real fix.

The usual pile of upstream fixes have made a lot of Gentoo packages
build more efficiently.  Many GNOME packages are now supported.