Update Gentoo to support Xfce PulseAudio plugins
The meta-package was updated to include PulseAudio plugins, so this
fixes installing related packages.

A workaround was added to skip BLKDISCARD in systemd-repart until
there is a real fix for Linux 5.10.  This fixes the Chromebook
example system's self-formatting after Gentoo dropped Linux 5.9.

The Firefox ports to unsupported architectures (i586 and PowerPC)
have been dropped.  It was a fun experiment to keep it running on
old CPUs, but it's not realistically usable on these systems with
under 1GiB of RAM, so it's time to lessen the maintenance burden.

Many Gentoo packages have been fixed upstream to drop unnecessary
dependencies and support cross-compiling in general.  Most GNOME
packages up to the introspection wall are now supported.  The Xfce
desktop environments in the example systems now enable just about
all optional features.