Switch initrds to zstd compression
Enough distros have migrated to Linux 5.10 and enabled zstd initrd
support, so make that the default format.  The only problematic
distros are Ubuntu and CentOS (and the EOL Fedora releases) because
their kernels are too old, which requires them to continue with xz.

The netbook example now produces a self-formatting GPT image like
the Chromebook example, so that systemd-repart is being tested on a
non-verity system.  Its final disk layout is just a boot partition
and two root partitions, with the rest used for persistent /var.

Gentoo has had a bunch of fixes applied upstream to increase build
efficiency and support more features.  In particular, binutils has
been eliminated from target system dependencies with GCC 10.