Use the Gentoo dist-kernel for building generic images
When building a bootable Gentoo system without providing a kernel
configuration, it will use a preconfigured distro kernel (which is
based on Fedora configs) automatically.  Providing a system kernel
config will continue to use the previous behavior of building an
unpackaged kernel.  The Gentoo-specific option raw_kernel can be
defined explicitly to override this behavior to force building an
unpackaged kernel or not regardless of provided configs.

Like with other distros, this kernel and its modules will not be
installed into the image by default to keep it minimal.  Add
sys-kernel/gentoo-kernel to packages to install them.  This is not
necessary just to boot the system since configuration options are
overridden to move modules to builtins for required functionality.

Note that no initrd is generated or required, so early microcode
updates are not included for applicable platforms by default.  This
might be changed to support Dracut initrds in the future, but for
now, skipping this step allows creating images for non-native CPUs
without juggling QEMU user binaries and binfmt configuration.

There were many efficiency improvements applied in various areas.
The package manager will no longer wait for files to sync to disk
after installing each package (which wasn't useful here because the
final product is a different file system image).  A bunch of unused
package builds were eliminated, like gettext and Git.