Support UEFI builds for non-x86_64 platforms by default
Previously, everything used x86_64 architecture names for files
related to UEFI, so the system configuration files needed to edit
them for different CPUs.  Now it will map the target architectures
to the appropriate file names so other CPUs work out of the box.

Gentoo has fixed cross-compiling the UEFI stub, so it will produce
unified kernels for any supported platform (mostly--e.g. arm and
arm64 support the UEFI stub, but binutils doesn't support the UEFI
format for objcopy to create the final image).  Since it supports
non-native UEFI images now, it needs a cross-compiled systemd build
with the stub enabled.  It will rebuild one if UEFI support was not
configured in the normal systemd package, which will be a potential
increase in build times (for the benefit of space saved from not
having the UEFI files packaged in the file system).

There were a handful of other Gentoo fixes applied upstream, like
eliminating unused builds of GNOME-related C++ bindings when those
packages are installed.