Replace PulseAudio with PipeWire in Gentoo
This configures PipeWire for socket-activation in user sessions if
it is installed, and otherwise falls back to PulseAudio.  Gentoo's
PulseAudio package does not differentiate client libraries/programs
from the server, so the regular PulseAudio server will be installed
alongside PipeWire and left unstarted.

Everything in the example systems should be fixed upstream to work
with GCC 10 at this point, which is now the default compiler on the
main architectures.  This includes a few efficiency improvements
like no longer building binutils and skipping unused tests.