Sync with upstream Gentoo fixes
A bunch of fixes have been applied upstream, so their conflicting
workarounds can be dropped here.  Many of these changes drop
unnecessary package builds and remove bottlenecks in the depgraph.

Since Python is now built with SQLite support, Firefox doesn't need
to have a specially configured stage3 anymore, so it can just be
installed as a regular package without manual changes.  To support
this, the latest versions of NSPR and NSS will always be accepted,
even when Firefox is not installed.