Add Fedora 34
I decided to move the older supported Fedora release to the legacy
directory, so the main fedora.sh only implements the newest release
version.  This will simplify things in the future when there are
required changes in new versions.

The README will now only advertize the latest two Ubuntu releases,
which is how Fedora is handled.  Older versions will continue to be
usable with no priority on keeping them fully functional.  (The
20.04 release was already out of date due to not having erofs-utils
new enough to support exclude lists.)