Fix a bunch of Gentoo build issues
Gentoo stabilized a lot of new versions with plenty of build
failures, so this makes required adjustments in addition to fixes
pushed upstream.  In particular, the previous colord version has to
be downloaded separately now, but this should be fixed soon.

The command-line syntax for setting arbitrary options now supports
omitting the equals sign.  This sets the option's value to its name
as a shorthand way of setting boolean options that only check if it
has a nonempty value, e.g. "install.sh -o networkd".

The FTL example container now implements the "nvidia" option to
support the proprietary driver to be consistent with all the other
Fedora example builds.  (The remaining two, BoI and VVVVVV, don't
need special handling to work on a host with proprietary drivers.)