Update Gentoo for Linux 5.13
There are only some minor config changes.  Also, the RISC-V example
has been restricted to 5.12 for now, since 5.13 fails to boot via
UEFI.  It could just replace the -bios option with -kernel in the
QEMU command to work with it, but UEFI testing was one of the main
reasons for the system in the first place.