Create an overlay for Gentoo fixes
Instead of editing things directly in the Gentoo package tree, any
modified files will be copied into a dedicated overlay which will
get priority.  The eclass changes still need to be made in the
Gentoo repository so they take effect across all packages.  The
method of copying and editing (as opposed to just using prewritten
files) should theoretically keep getting updates from upstream, but
they might push conflicts that require manual intervention.

An avalanche of broken commits was pushed to both testing and
stable packages in Gentoo, so this includes significant changes to
deal with the fallout.
  - A circular dependency was added in pango which requires
    bootstrapping in the native buildroot.  (This dependency should
    probably just be dropped in the overlay, but I left it for now
    assuming it is useful for the target systems.)  The package was
    not compatible with the cross-multilib setup, so remaining EAPI
    issues were addressed in the overlay, and the multilib option
    has been removed from the example systems.
  - The Linux kernel headers were added for 5.13 which break clang
    sanitizers, so they're pinned to 5.12.
  - SELinux, autotools, and other core things were broken, and they
    are all now fixed in the overlay.

ARM's hard-float ABI can now be specified with the Gentoo-specific
boolean option hardfp.  It defaults to enabled on armv6 and armv7.
This allows setting the ABI without needing to override the host
and Rust target values.  The option also controls which stage3 is
downloaded for the native build system, but this shouldn't be an
issue unless a slow 32-bit ARM chip is being used to cross-compile
for a different ABI.

The stage3s have been updated to use hardened OpenRC variants for
x86 arches, which are the new ones.  This changed the URL format.

USE flags for introspection and its user vala are now disabled by
default.  Introspection can't be cross-compiled sensibly, and it is
unlikely to be used on the build system, so this saves build time.
A few other minor unnecessary package builds have been skipped to
save additional time, and the unusable eselect packages have been
removed from the target systems.  Firefox 90 has dropped gtk2, so
all of its related packages should no longer be built for most
systems.  Overall build times are getter faster for large systems.